backend/requirements.txt
fastapi
uvicorn
pydantic
huggingface_hub
python-dotenv
networkx
requests
watchfiles
